updated files

+++ contracts/interface/I2Pool.sol
@@ -3,6 +3,13 @@ pragma solidity ^0.8.9;
 
 interface I2Pool {
     function coins(uint256 arg0) external returns (address);
-    function calc_token_amount(uint[2] calldata amounts, bool _h) external returns (uint);
-    function add_liquidity(uint256[2] memory _deposit_amounts, uint256 _min_mint_amount) external returns (uint256);
+
+    function calc_token_amount(uint256[2] calldata amounts, bool _h)
+        external
+        returns (uint256);
+
+    function add_liquidity(
+        uint256[2] memory _deposit_amounts,
+        uint256 _min_mint_amount
+    ) external returns (uint256);
 }
