+++ contracts/interface/ISushiSwapRouter.sol
@@ -3,15 +3,15 @@ pragma solidity ^0.8.9;
 
 interface ISushiSwapRouter {
     function swapExactTokensForTokens(
-        uint amountIn,
-        uint amountOutMin,
+        uint256 amountIn,
+        uint256 amountOutMin,
         address[] calldata path,
         address to,
-        uint deadline
+        uint256 deadline
     ) external;
 
-    function getAmountsOut(uint amountIn, address[] memory path)
+    function getAmountsOut(uint256 amountIn, address[] memory path)
         external
         view
-        returns (uint[] memory amounts);
+        returns (uint256[] memory amounts);
 }
